changed RUs type to float64 and changed parsing method

+++ cosmosapi/query.go
@@ -16,7 +16,7 @@ type QueryDocumentsResponse struct {
 	Documents    interface{}
 	Count        int `json:"_count"`
 	Continuation string
-	RUs          int
+	RUs          float64
 }
 
 // NewQuery create a query with given parameters.
